+++ src/org/francis/intel/challenge/search/SearchUtils.java
@@ -2,20 +2,20 @@ package org.francis.intel.challenge.search;
 
 public class SearchUtils implements Constants {
 
-    public static byte complementDir(int dir) {
-        return (byte)(dir^1);
+    public static int complementDir(int dir) {
+        return (dir^1);
     }
     
-    public static byte forbidDir(byte dir) {
-        return (byte)(1 << (dir+LEFT_SHIFT_OFFSET));
+    public static int forbidDir(int dir) {
+        return (1 << (dir+LEFT_SHIFT_OFFSET));
     }
     
-    public static byte allowOnlyDir(byte dir) {
-        byte forbidMask = (byte)(1 << (dir+LEFT_SHIFT_OFFSET));
-        return (byte)(forbidMask ^ MASK_CONSTRS);
+    public static int allowOnlyDir(int dir) {
+        int forbidMask = (int)(1 << (dir+LEFT_SHIFT_OFFSET));
+        return (forbidMask ^ MASK_CONSTRS);
     }
     
-    public static int nxtPos(int pos, byte dir, int sPos, int width, int totalLength) {
+    public static int nxtPos(int pos, int dir, int sPos, int width, int totalLength) {
         switch (dir) {
             case UP :
                 return pos-width;
